Optimized imports everywhere //TODO make tests passed

+++ app/src/main/java/com/example/weatherforecast/data/db/DailyWeatherEntity.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "daily_weather")
 data class DailyWeatherEntity(
-    @PrimaryKey(autoGenerate = true) val id: Long = 0L,
+    @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val date: String,
     val dt: Long,
     val temp: Double,
